bgp_asn_extended

+++ internal/service/ec2/vpnsite_customer_gateway.go
@@ -116,7 +116,7 @@ func resourceCustomerGatewayCreate(ctx context.Context, d *schema.ResourceData, 
 			return sdkdiag.AppendFromErr(diags, err)
 		}
 
-		input.BgpAsnExtended = aws.Int64(int64(v))
+		input.BgpAsnExtended = aws.Int64(v)
 	}
 
 	if v, ok := d.GetOk(names.AttrCertificateARN); ok {
